// plugwerk-server/plugwerk-server-frontend/src/api/generated/model/review-item-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Plugwerk API
 * Plugin marketplace for the Java/PF4J ecosystem
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ReviewItemDto
 */
export interface ReviewItemDto {
    /**
     * 
     * @type {string}
     * @memberof ReviewItemDto
     */
    'releaseId': string;
    /**
     * 
     * @type {string}
     * @memberof ReviewItemDto
     */
    'pluginId': string;
    /**
     * 
     * @type {string}
     * @memberof ReviewItemDto
     */
    'pluginName': string;
    /**
     * 
     * @type {string}
     * @memberof ReviewItemDto
     */
    'version': string;
    /**
     * 
     * @type {string}
     * @memberof ReviewItemDto
     */
    'submittedBy'?: string;
    /**
     * 
     * @type {string}
     * @memberof ReviewItemDto
     */
    'submittedAt': string;
    /**
     * 
     * @type {string}
     * @memberof ReviewItemDto
     */
    'artifactSha256'?: string;
}